Baseline model (#13)
* Added first version of validation model one. AB#48755

Because:
- Needed for validating model

This is the first version. Optimizing is still required.

* Created first version of the baseline model.

Because:
 - It gives a minimal score.

First version of the baseline model using tensorflow keras Functional API.

* Added second version of baseline model. AB#49514

Because:
- Needed for the model.

This is the second version. Optimizing is still required.

* Updated code in baseline_model.py AB#49514

Because:
- It needs to be more reuseable
- It needed to look better

Changed function names and variables. Edited model layout

* Updated docstrings for baseline_model.py AB#49514

Because:
- Flake8 linting did not approve
- Functions needed better documentation

Updated dosctrings for all functions so it is approved by flake8 linting.

* Added layer to baseline_model.py AB#49514

Because:
- It missed one

Added a layer (kernel size 3, filters 64) to baseline model.

* Added first version of analyse for the baseline model also edited the model.

* Fixed the imports for the baseline model.

* Changed docstrings and added typehinting

Because:
- Docstrings needed to be better
- Paramters in all functions didn't have a type

Edited docstrings and added typehinting to for all functions in baseline_model.py

* Small changes in the type of hinting for the functions.

* Added baseline model requirements to requirements.txt

* Deleted validation model form the baseline model.

* Added compile and train model part to notebook

Because:
- It wasn't finished yet

Added the part "model compile and train" to the notebook

* Made small change in read_data().

Because:
- An error occurred

Edited json.loads(data) to json.load(data.read())

* Added an callback to train_model()

Because:
- It is needed to plot the training results

Added an Histroy callback from keras.callbacks so it can be used to plot the results

* Added test function to baseline_model.py and edited notebook

Because:
- Needed for further testing of the models accuracy
- Notebook needed to be changed based on these changes

Added 1 function to baseline_model.py (test_model). Changed create_trainset to create_datasets. Added cells to notebook.

* Changed function name and edited docstring in baseline_model.py

Because:
- Function name wasn't good
- Linting errors

Changed test_model to evaluate_model and edited the docstring of the function.

* Changed model structure so it won't overfit.

Because:
- previous version sucked - from brandon

* Finished baseline model (documentation and code) AB#49514

Because:
- It took a long time.
- It's the first version which is necessary.
- It needed good documentation.

Finished baseline model that can check if someone is happy or not.

* forgot to commit settings.json.example

Co-authored-by: Maria <62339103+XXX-MA@users.noreply.github.com>
Co-authored-by: stormjoannes <stormjoannes10@hotmail.com>