Baseline model (#16)
* Added first version of validation model one. AB#48755

Because:
- Needed for validating model

This is the first version. Optimizing is still required.

* Created first version of the baseline model.

Because:
 - It gives a minimal score.

First version of the baseline model using tensorflow keras Functional API.

* Added second version of baseline model. AB#49514

Because:
- Needed for the model.

This is the second version. Optimizing is still required.

* Updated code in baseline_model.py AB#49514

Because:
- It needs to be more reuseable
- It needed to look better

Changed function names and variables. Edited model layout

* Updated docstrings for baseline_model.py AB#49514

Because:
- Flake8 linting did not approve
- Functions needed better documentation

Updated dosctrings for all functions so it is approved by flake8 linting.

* Added layer to baseline_model.py AB#49514

Because:
- It missed one

Added a layer (kernel size 3, filters 64) to baseline model.